refactor: 게임판 생성 기능 Board 내부로 이동
