+++ README.md
@@ -4,10 +4,10 @@
 
 ### 기능 목록
 
-- [ ] 지뢰찾기 게임판(Board)을 그린다.
-- [ ] 게임판은 포지션과 마크를 갖는다.
-    - [ ] 게임판은 2차원 칸을 갖는다.
-    - [ ] 게임판은 지뢰 위치를 받아 지뢰를 배치한다.
+- [X] 지뢰찾기 게임판(Board)을 그린다.
+- [X] 게임판은 포지션과 마크를 갖는다.
+    - [X] 게임판은 2차원 칸을 갖는다.
+    - [X] 게임판은 지뢰 위치를 받아 지뢰를 배치한다.
 - [X] 위치 생성기는 최대 x, y 좌표를 받아 랜덤한 위치를 생성한다.
 - [X] 마크는 지뢰와 일반 칸이 존재한다
 - [X] 포지션은 x, y 좌표를 갖는다.
