+++ src/main/kotlin/minesweeper/domain/Board.kt
@@ -3,13 +3,22 @@ package minesweeper.domain
 import minesweeper.domain.field.Mark
 import minesweeper.domain.field.Position
 
-class Board(val board: Map<Position, Mark>) {
-
-    fun plantMines(minePositions: Set<Position>): Board {
-        val withMineBoard = board.toMutableMap()
-        minePositions.forEach { position ->
-            withMineBoard[position] = Mark.MINE
+class Board private constructor(val markMap: Map<Position, Mark>) {
+    companion object {
+        fun create(width: Int, height: Int, minePositions: Set<Position>): Board {
+            val board = generateAllPositions(width, height)
+                .associateWith { Mark.SAFE }
+                .toMutableMap()
+            board.putAll(minePositions.associateWith { Mark.MINE })
+            return Board(board.toMap())
         }
-        return Board(withMineBoard.toMap())
+
+        private fun generateAllPositions(width: Int, height: Int): List<Position> {
+            return (0 until width).flatMap { i ->
+                (0 until height).map { j ->
+                    Position(i, j)
+                }
+            }
+        }
     }
 }
